+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### API Changes
 - `GmGM` now defaults to `centering_method=None`.
 - Added argument `GmGM(_assume_sparse=False)`, which when True forces `GmGM` to treat the input as a sparse matrix.  This is for use when the input is sparse but cannot be detected by `scipy.sparse.issparse()`, such as `AnnData`'s experimental sparse backed formats.
+- If `GmGM(threshold_method='overall', to_keep: MaybeDict[int])` or `GmGM(threshold_method='overall-col-weighted', to_keep: MaybeDict[int])`, no longer throws an error but rather converts `to_keep` to `to_keep / axis_size` and proceeds from there.
 
 ### Improvements
 - Added `overall-col-weighted` thresholding method